+++ include/instructions/RabbitizerOperandType.h
@@ -26,6 +26,22 @@ typedef enum RabbitizerOperandType {
 #undef RAB_DEF_OPERAND
 
 
+struct RabbitizerInstruction;
+
+typedef size_t (*OperandCallback)(const struct RabbitizerInstruction *self, char *dst, const char *immOverride, size_t immOverrideLength);
+
+extern const OperandCallback instrOpercandCallbacks[];
+
+
+#define RAB_DEF_OPERAND(prefix, operand) size_t RabbitizerOperandType_process_##prefix##_##operand (const struct RabbitizerInstruction *self, char *dst, const char *immOverride, size_t immOverrideLength);
+
+#include "instructions/operands/RabbitizerOperandType_cpu.inc"
+#include "instructions/operands/RabbitizerOperandType_rsp.inc"
+#include "instructions/operands/RabbitizerOperandType_r5900.inc"
+
+#undef RAB_DEF_OPERAND
+
+
 #ifdef __cplusplus
 }
 #endif
